chore: add packages/shared and centralize roles/supabase
