+++ apps/web/package.json
@@ -23,7 +23,8 @@
     "lucide-react": "^0.544.0",
     "next": "15.2.8",
     "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "react-dom": "19.1.0",
+    "@mysryear/shared": "0.0.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
