+++ package.json
@@ -12,8 +12,8 @@
     "lint": "npm --workspace apps/web run lint",
     "format": "prettier --write --ignore-unknown .",
     "format:check": "prettier --check --ignore-unknown .",
-    "typecheck": "npm --workspace apps/web run typecheck",
-    "test": "npm --workspace apps/web run test",
+    "typecheck": "npm --workspace packages/shared run typecheck && npm --workspace apps/web run typecheck",
+    "test": "npm --workspace packages/shared run test && npm --workspace apps/web run test",
     "test:e2e": "npm --workspace apps/web run test:e2e",
     "verify": "npm run lint && npm run typecheck && npm run test && npm run build"
   },
