+++ packages/shared/src/index.ts
@@ -1,0 +1,6 @@
+export type { UserRole } from './types/roles'
+export { USER_ROLES } from './types/roles'
+
+export { getSupabaseEnv } from './supabase/env'
+export { createWebSupabaseClient } from './supabase/browser'
+export { createWebSupabaseServerClient, type SupabaseCookieMethods } from './supabase/server'
