+++ packages/shared/src/supabase/env.ts
@@ -1,0 +1,14 @@
+function requireEnv(name: string) {
+  const value = process.env[name]
+  if (!value) {
+    throw new Error(`Missing required env var: ${name}`)
+  }
+  return value
+}
+
+export function getSupabaseEnv() {
+  return {
+    url: requireEnv('NEXT_PUBLIC_SUPABASE_URL'),
+    anonKey: requireEnv('NEXT_PUBLIC_SUPABASE_ANON_KEY'),
+  }
+}
